+++ client/src/pages/Customers.tsx
@@ -16,6 +16,7 @@ import VisibilityIcon from "@mui/icons-material/Visibility";
 import IconButton from "@mui/material/IconButton";
 import Tooltip from "@mui/material/Tooltip";
 import DeleteIcon from "@mui/icons-material/Delete";
+import Swal from "sweetalert2";
 
 import { getCustomers, deleteCustomer } from "../api/customers";
 import { CustomersResponseType } from "../types/customers";
@@ -84,15 +85,26 @@ const Customers = () => {
   };
 
   const handleDeleteCustomer = async (cid: string) => {
-    try {
-      setLoading(true);
-      await deleteCustomer(cid);
-      const customers = await getCustomers();
-      setData(customers);
-      setLoading(false);
-    } catch (er) {
-      setLoading(false);
-    }
+    Swal.fire({
+      title: "系統訊息",
+      text: "您確定要刪除此筆資料?(資料刪除後將無法復原)",
+      showCancelButton: true,
+      cancelButtonText: "取消",
+      confirmButtonText: "確定",
+      confirmButtonColor: "#3f50b5",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          setLoading(true);
+          await deleteCustomer(cid);
+          const customers = await getCustomers();
+          setData(customers);
+          setLoading(false);
+        } catch (er) {
+          setLoading(false);
+        }
+      }
+    });
   };
 
   return (
